requirements.txt
pip==25.3
PyYAML==6.0.2
pyTelegramBotAPI==4.29.1
pydantic==2.12.5
pydantic_core==2.41.5
schedule==1.2.2
requests==2.32.5